make hashlib binary safe.
this patch fixes hashlib which does not respect Buffers as it should.

The macro converts argument positions to raw data. I was forced to use
it because i wasn't able get the value out of scope of a lookup function
withcout beeing collected by the garbage collector. It could be
optimized a little bit, but I didn't found a way to just declare a
Utf8Value and initialize it later.